Fix Node.path cacheing bug
This was a tough bug to figure out --- mysterious '.html' strings were
being appended to node slugs.

Turns out when I added cacheing to the Node.path property I left it
returning the same individual list object every time it was called. This
object was then being modified in-place by the Page.get_filepath()
method which assumed it was dealing with a disposable copy.

For future robustness Node.path now always returns a disposable copy of
the cached path list.