feature: add config to suppress generation of index.html files for directories
If the config file contains an entry `emit_no_file_directory_nodes`
whose value is `False`, the new function `_generate_html()` in
`ark/nodes.py` will return `False`, which will in turn make
`Node.write()` exit without generating an `index.html` page for a
directory. We cannot rely on the `load_node_dir` filter for this
because when that filter's value is `False`, Ark does not recurse into
the directory at all, which means that files in child directories
aren't translated.

closes #21

name